New Columns added.
This is the same as "get_stock_price.py" except for two new columns. One titled "DailyDiff" which is the raw difference between the opening and closing price of one day for a specific symbol. The other is titled "Bull/Bear" which performs a binary classification on "DailyDiff" using a sigmoid function. Bull is 1. Bear is 0.